Trac Ticket #290 - server hangs during shutdown if betxn pre/post op fails
https://fedorahosted.org/389/ticket/290

Fix description: The logic to call dblayer_txn_abort in the
ldbm_back update functions was wrong.  If the operations fail,
the abort function has to be called if dblayer_txn_begin is
already called and dblayer_txn_commit is not.  This patch
checks the txn value set by dblayer_txn_commit to determine
to call dblayer_txn_abort or not.