ticket 304 - Fix kernel version checking in dsktune
The kernel version cehcking in dsktune is incorrect.  It takes each
version element and checks it separately, without factoring in the
check for the previous version element.  This causes checks like
3.2.5 > 2.4.7 to be evaluated as false.

This fix factors in the previous version element comparison, which
ensures that the version is evaluated correctly as a whole.
(cherry picked from commit e2a1090098fd2bcdd37dc1e9870cad46d975761c)