Ticket #320 - allow most plugins to be betxn plugins
This allows plugins to run as betxn plugins instead of regular pre/post
op plugins.  For plugins that are not preoperation or postoperation in
the DSE, e.g. that are "object", I've introduced a boolean attribute
nsslapd-pluginbetxn.  If this is set to a true value, the internal
pre/post operation plugins registered will be made betxn plugins
instead.
Looks like cos cannot currently be made a betxn plugin - there is a deadlock
when deleting from the ldbm and DSE backends at the same time.
Everything else passes the acceptance tests.
Reviewed by: nhosoi (Thanks!)