Bug 804828 - clear-password mod from replica is     discarded before changelogged
https://bugzilla.redhat.com/show_bug.cgi?id=804828

Description:
unhashed password pseudo-attribute is necessary for plugins that
handle passwords.  The plugin could be any one including pre/post-
plugin and be-txn-pre/post-plugin.  On the other hand, the pseudo-
attribute should not be in the database.

This patch declares the unhashed password pseudo-attribute as an
operational attribute and allows the pseudo-attribute in an entry
in the memory.  When creating a flat entry by entry2str_ function
to store in the database, it omits the pseudo-attribute.
(backported commit c06a8faa9140668836dcf02722908319ec2e6e1e)