Trac Ticket #46 - (take 3) setup-ds-admin.pl does not like ipv6 only hostnames
https://fedorahosted.org/389/ticket/46

Fix Description:
Perl functions (gethostbyname, gethostbyaddr) do not support IPv6
addresses.  This patch addes the IPv6 support if an IPv6 enabled
socket module is available.  If Socket::addrinfo subsystem is
found, use getaddrinfo and getnameinfo from the module. If not
and if Socket6 module is available, the module is used.  If both
are not found on the system, the existing non-IPv6 methods are
called.