Ticket #337 - RFE - Improve CLEANRUV functionality
Bug Description:  Previously the steps to remove a replica and its RUV was problematic.
                  I created two new "tasks" to take care of the entire replication environment.

Fix Description:

[1] The new task "CLEANALLRUV<rid>" - run it once on any master

    This marks the rid as invalid. Used to reject updates to the changelog, and the database RUV
    It then sends a "CLEANRUV" extended operation to each agreement.
    Then it cleans its own RUV.

    The CLEANRUV extended op then triggers that replica to send the the same CLEANRUV extop to its replicas, then it cleans its own RID. Basically
this operation cascades through the entire replication environment.

[2] The "RELEASERUV<rid>" task - run it once on any master

    Once the RUV's have been cleaned on all the replicas, you need to "release" the rid so that it can be reused. This operation also cascades thro
ugh the entire replication environment. This also triggers changelog trimming.

For all of this to work correctly, there is a list of steps that needs to be followed. This procedure is attached to the ticket.

https://fedorahosted.org/389/ticket/337

reviewed by: ?
(cherry picked from commit 0f50544b9567907edd0ba645951d7cd325354107)