Bug #361: Bad DNs in ACIs can segfault ns-slapd
A bad userdn will make the DN normalization fail.  This sets dn to NULL,
and the server will attempt to access the NULL dn.
The fix is to just return with an error if the normalization fails since
the DN is invalid.
Reviewed by: rmeggins
(cherry picked from commit 8ba2982457e7d2143ae0c2af212412122592c284)