Bug 832560 - slapi_attr_value_cmp: Process /usr/sbin/ns-slapd was killed by signal 11 (SIGSEGV)
https://bugzilla.redhat.com/show_bug.cgi?id=832560

Bug Description: When a conflict occurred and an attribute to add
in the mods is deleted by another master, the conflict checking
code failed to detect it.

Fix Description: Addng the NULL check to the conflict checking
code.  If it is found NULL, the conflict check is skipped and the
index update call is skipped, as well.