improve txn test index handling
Sometimes it is difficult to determine which indexes actually have files,
and the txn thread would keep looping forever looking for missing files.
This changes the txn thread to retry a few times, then just start skipping
missing indexes.  The txn thread will log which indexes it uses and which
it skipped.