Trac Ticket #402 - nhashed#user#password in entry extension
https://fedorahosted.org/389/ticket/402

Fix description: This patch adds the method to use entry
extension to stash the unhashed password in addition to the
existing one which uses the ordinary attribute.

It introduces the definition "USE_OLD_UNHASHED" in configure.ac
to keep the old method to use the attribute.

Once all the plugins' migration is done, the old method can be
disabled by removing the definition.  We could also remove the
code in "#if defined(USE_OLD_UNHASHED)" then.

The first proposal was reviewed and commented by nkinder.
(Regarding the comments, see also the trac ticket.  Thanks
a lot, Nathan!) This second patch includes the fixes pointed
out by him.