Ticket #410 - Referential integrity plug-in does not work when update interval is not zero
https://fedorahosted.org/389/ticket/410
Resolves: Ticket #410
Bug Description: Referential integrity plug-in does not work when update interval is not zero
Reviewed by: nhosoi (Thanks!)
Branch: 389-ds-base-1.2.11
Fix Description: Checking for newsuperior == NULL is not sufficient - also
have to check for slapi_sdn_get_dn(newsuperior) == NULL as well.  Otherwise
we write the string "(null)" to the referint log and that blows up the code
that attempt to apply the change.
Platforms tested: RHEL6 x86_64
Flag Day: no
Doc impact: no