Ticket 403 - CLEANALLRUV feature
This is a rewrite of the first version of CLEANALLRUV.  The task is now an official slapi task, but the
older/original method still works as well.  You can run multiple tasks(4) at the same time.  The task can handle
situations where a replica is down, or "older"replicas are in the topology.  I've also added an "abort"
task to cancel a particular cleanallruv task, as the task can potentially run for a long time while
waiting for a server to come online or be cleaned.

https://fedorahosted.org/389/ticket/403

Reviewed by: richm & noriko (Thanks!)
(cherry picked from commit 657efc69487cc1b04d82642dddd47f93b7a37cf1)