Ticket #426 - support posix schema for user and group sync
https://fedorahosted.org/389/ticket/426
Resolves: Ticket #426
Bug Description: support posix schema for user and group sync
Reviewed by: nhosoi, mreynolds (Thanks!)
Branch: 389-ds-base-1.2.11
Fix Description: This is a new winsync plugin.  By default it works with the
posixaccount and posixgroup schema included with user and group entries in
Windows 2003 R2 and later Active Directory.  There is a switch that can be
set to use the older MSSFU schema.
This is not entirely bi-directional - when you add new entries to DS, the
posix attributes in those new user/group entries are not synced to AD -
this is due to a limitation in the old winsync api v1.  Future work is
planned to add that capability.
This was copied from the upstream at
https://github.com/cgrzemba/Posix-Winsync-Plugin-for-389-directory-server
with permission from the author.
Platforms tested: RHEL6 x86_64, Windows 2008 R2
Flag Day: no
Doc impact: no