Trac Ticket #340 - Change on SLAPI_MODRDN_NEWSUPERIOR is not       evaluated in acl
https://fedorahosted.org/389/ticket/340

Bug Description: When modrdn operation was executed, only newrdn
change was passed to the acl plugin.  Also, the change was used
only for the acl search, but not for the acl target in the items
in the acl cache.

Fix Description: This patch also passes the newsuperior update
to the acl plugin.  And the modrdn updates are applied to the
acl target in the acl cache.