Trac Ticket #340 - Change on SLAPI_MODRDN_NEWSUPERIOR is not       evaluated in acl
https://fedorahosted.org/389/ticket/340

Bug Description: When modrdn operation was executed, only newrdn
change was passed to the acl plugin.  Also, the change was used
only for the acl search, but not for the acl target in the items
in the acl cache.

Fix Description: This patch also passes the newsuperior update
to the acl plugin.  And the modrdn updates are applied to the
acl target in the acl cache.
(cherry picked from commit 5beb93d42efb807838c09c5fab898876876f8d09)
1.2.10 - cherry-picked from 7399cbd53d6289df592d3414a84972eacb4dc97d
- had to fix a couple of conflicts in the acl code