Trac Ticket #466 - entry_apply_mod - ADD: Failed to set     unhashed#user#password to extension
https://fedorahosted.org/389/ticket/466

Bug Description: If changes are stuck in the chagnelog and the
server is restarted, the changes are being replicated as soon
as plugin_startall is called.  At the point, the entry extension
has to be already initialized.

Fix Description: The entry extension initialization which used
to be called in pw_init.  Now it is moved to just before
plugin_startall in main.