Trac Ticket #470 - 389 prevents from adding a posixaccount     with userpassword after schema reload
https://fedorahosted.org/389/ticket/470

Bug description: Schema reload task reloads schema files in the
schema directory.  Not just them, DS has several internal schema
which are not stored in the schema file, which were lost after
the schema reload task is executed.  One of them unhashed#
user#password was necessary for adding a posixaccount.

Fix description: When registering an internal schema, the schema
is stashed in a hash table.  When schema reload is executed, the
internal schema are reloaded with the external schema.