Trac Ticket #451 - Allow db2ldif to be quiet
https://fedorahosted.org/389/ticket/451

Description: Export utility db2ldif and import utility ldif2db are
always verbose.  This patch is adding an option '-q' to suppress
the outputs.

This patch modifies db2bak and bak2db which already supports '-q',
as well.  The scripts used to interpret '-q'.  If it was not set,
a verbose debug option '-d 524288' was passed to ns-slapd.  Instead,
this patch directly passes '-q' to ns-slapd and let it handle it
in the C code.