Trac Ticket #276 - Multiple threads simultaneously working on     connection's private buffer causes ns-slapd to abort
https://fedorahosted.org/389/ticket/276

Bug description: When a connection is to be released, the current
code releases the connection object before making it readable,
which leaves a small window for multiple threads accessing the
same private buffer.

Fix description: This patch moves the location of releasing the
connection object after the connection is readable.