Trac Ticket #494 - slapd entered to infinite loop during new index addition
https://fedorahosted.org/389/ticket/494

Bug Description: An entry which rdn is nsuniqueid=...,nsuniqueid=...
was accidentally generated.  It broke the entryrdn tree and put
the reindex and export into the the infinite loop.

Fix Description: Added an missing check if the retrieved entry is
already an tombstone or not.  This tombstone is the one converted
by the entryusn plugin.  Please note that replication has an urp
code which handles such conflicts.