Bug 866623 - RHDS crash when attempting to free an invalid pointer
https://bugzilla.redhat.com/show_bug.cgi?id=866623

Bug Description: When a simple-paged-results search is abandoned,
it calls pagedresults_cleanup and cleans up the search result set
stashed in the connection object.  The abandoned fact is not stored
any place.  Thus if the further operation expecting the search
result set is issued, it crould crash the server.

Fix Description: This patch introduces ABANDONED flag and sets it
if the simple paged results search is abandoned.  Any further ops
on the search are skipped.

This patch is also adding a check if the given page size is 0 in
the simple-paged-results control or not.  If it is 0, treat is as
an abandoned operation.