Ticket #503 - Improve AD version in winsync log message
When enabling replication level logging, winsync prints out what
version of Windows/AD it detects.  If it detects win2k3 or later, it
prints out "detected win2k3 peer".  This can be confusing if you are
running a later version, such as win2k8.  The code is really trying
to detect if we can use certain operations that only started being
supported in win2k3.  This patch changes the message to match this
logic by printing "detected win2k3 or later peer".