Trac Ticket #311 - IP lookup failing with multiple DNS     entries
https://fedorahosted.org/389/ticket/311

Bug description: DNS keyword in ACI only accepted an FQDN returned
from gethostbyaddr.  If an alias hostname was set in an ACI, a request
sent from the host was treated as the one from the primary hostname
and it failed to get the expected access rights.

Fix description: This patch turns the "dns" keyword covers the alias
hostnames, as well.  In addition to the primary hostname, by setting
the secondary hostnames as dns, clients requests would obtain the
expected access rights.  When an IP address is associated with multiple
hostnames (primary: hostA, aliases: hostB and hostC), they could be
listed, for instance, in an aci as follows:
aci: (targetattr = "*") (version 3.0;acl "dns example";allow (all)
  dns="hostA.example.com" or dns="hostB.example.com" or dns="hostC.
  example.com";)