Trac Ticket #520 - RedHat Directory Server crashes (segfaults) when moving ldap entry
https://fedorahosted.org/389/ticket/520

Fix description: The code to check if the new superior entry exists
or not was returning the "No such object" error only when the op
was requested by the directory manager.  This patch is removing
the condition so that whoever the requester is, it returns the error.
(cherry picked from commit 40e68b18dae23a21654719a742419f172aa54806)