Trac Ticket #499 - Handling URP results is not corrrect
https://fedorahosted.org/389/ticket/499

Bug description: When an urp resolution occurred as follows
  [] - urp_delete: Entry "nsuniqueid=<UNIQID>,uid=<UID>,o=<ORG>"
is already a Tombstone.
the operation should be skipped in the backend, but should return
SUCCESS to the supplier.  Otherwise, the supplier continues to
send the relay and the replication stops there.

Fix description: This patch introduced SLAPI_PLUGIN_NOOP (-2)
to the bepre and betxnpre plugin return value set (SLAPI_
PLUGIN_SUCCESS == 0; SLAPI_PLUGIN_FAILURE == -1).  If SLAPI_
PLUGIN_NOOP is returned, the backend code skips the operation,
but it returns SUCCESS.  Note that urp is only executed on the
replicated operation (not on the end user ones).