Ticket 549 - DNA plugin no longer reports additional info when range is depleted
When the DNA plug-in was modified to allocate range values at the
bepreop phase, it stopped returning detailed error strings to the
client when the range was depleted.

This patch allows our bepreop functions to fill in the error string
that the caller can return to the client.
(cherry picked from commit 3b177a768a48857f9dcb891a7dd3c8f8d9a220a4)