Ticket #545 - Segfault during initial LDIF import: str2entry_dupcheck()
Bug description: If an attribute type having multiple values exists
in an entry, and any compare function is not associated with the
attribute type, fast_dup_check flag was not disabled from the second
time.  Since fast_dup_check requires the compare function based on
the attribute syntax, it causes the segfault.

Fix description: This patch checks whether a compare function is
associated with the multi-valued attribute not just at the first
time the attribute type appears but at the second time and after,
and disable fast_dup_check properly if needed.

https://fedorahosted.org/389/ticket/545

Reviewed by Rich (Thank you!!)