Ticket #537 - Improvement of range search
Fix description: The index range search function index_range
_read_ext was written to call idl_fetch_ext to get an idlist
belonging to one key. Then add it to the main idlist as long
as the key satisfiles the range search filter condition.
This patch introduces a new range search function idl_new_
range_fetch to the new idl code, which generates an idlist
in one idl function that eliminates the redundancy such as
generating idlist and cursor per key.

This patch only implements the new idl version.  If idl_new
is not set, the existing code is executed.

Additionally, idl_new_fetch did not abort the read transaction
even if any error occurred in the transaction. Now, it switches
between commit and abort based upon the result.

https://fedorahosted.org/389/ticket/537

Reviewed by Rich (Thank you!!)